Switch RowVersionManager to using the FixedSizeAllocator (#19392)
This PR switches the `RowVersionManager` that tracks whether or not rows
are relevant for a given transaction to using the `FixedSizeAllocator`.
This makes the majority of its memory usage trackable with the buffer
manager. The `FixedSizeAllocator` does not support evicting yet - but it
could in theory also start supporting that in the future using this
setup.