Reuse correct table-level metadata during checkpoints (#20267)
We've noticed that checkpoints were occasionally reading a lot of
metadata (Gigabytes) when the `experimental_metadata_reuse`
functionality was enabled. The reason for this is that when an unchanged
table was finalized, `SingleFileTableDataWriter::FinalizeTable` would
reread and pin all of the metadata blocks that were written in a prior
checkpoint (not just for this table, but other tables too). The logic
that would do this in `SingleFileTableDataWriter::FinalizeTable` would
do:
```
MetadataReader reader(metadata_manager, pointer);
auto blocks = reader.GetRemainingBlocks();
metadata_manager.ClearModifiedBlocks(blocks);
```
As the metadata is written out by a single MetadataWriter for all
tables, this would potentially load and repin the metadata for all
tables (as `reader.GetRemainingBlocks()` would not know when to stop
reading).

The PR here changes the code in
`SingleFileTableDataWriter::FinalizeTable` to instead identify the
metadata blocks of the given table that are left unchanged (and keep
exactly those around).